requirements.txt
pdfplumber==0.10.2
pytesseract==0.3.10
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.3.0
tabula-py==2.9.0